requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested==1.2.24
drf-nested-routers==0.93.5
idna==3.7
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
routers==0.10.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
